Updates to Prepare Repo for Open Sourcing (#62) (#63)
