+++ README.md
@@ -1,8 +1,10 @@
 # Quadro
 
-Quadro is a kanban board to view your Github issues in the Deus_Lex repository.
+Quadro is a kanban board bootstrapped from the Create React App.
 
-It is currently hosted at [quadro-beta.stage](https://quadro-beta.stage.lexmachina.com/)
+Based on the current configuration in the `src/config` directory, it connects with the Github API (both GraphQL and REST) to pull in Github Issues.
+
+It is currently hosted at [https://quadro.lexmachina.com](https://quadro.lexmachina.com)
 
 ## Requirement
 
@@ -28,24 +30,28 @@ You will need to open two tabs in your terminal and be inside the project in bot
 
 Any work that you'd like to push to the repo, please do so in a branch off of `develop` via a pull request. Assign the PR to `@wired4code`. All PRs will be into `develop`, as deployment is handled via `master`.
 
+## Forking This Repo
+
+If you'd like to fork this repository and customize it to use with Github or another service, you will need to update `api.js` inside the `src/config` directory. This file contains all the API-related data, e.g., the API endpoints, owner name, repository, and customizations for the board.
+
 ## Available Scripts
 
 In the project directory, you can run:
 
-### `node server`
-
-This will start the Express server. This is needed to get the Oauth token to access Github.
-You will want to open up two terminals. In the one, run `node server`. In the other,
-run `npm run build`. Then go to `http://localhost:8080`.
+To run the application locally while developing, just run `npm run dev`. As long as you have the redirect_url above set to `http://localhost:3000/oauth/redirect`, `npm run dev` will be the only command you need to run.
 
 ### `npm run dev`
 
 Runs the app in the development mode.<br>
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser. You won't
-be able to interact with Github.
+Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
 
 The page will reload if you make edits.<br>
 You will also see any lint errors in the console.
+
+### `node server`
+
+This will start the Express server. To run the production build, run `node server` in one terminal. In the other,
+run `npm run build`. Then go to `http://localhost:8080`.
 
 ### `npm test`
