+++ src/components/Board.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { string, arrayOf, shape } from "prop-types";
 import '../App.scss';
 import CardContainer from './CardContainer';
-import { CONFIG, GET_BUCKET, getQueryString } from "../helpers/github";
+import { GET_BUCKET, getQueryString } from "../helpers/api_interface";
+import { CONFIG } from "../config/api";
 
 export default function Board( { member, statusLabels }) {
   const allQueryStrings = CONFIG.buckets.reduce((result, bucket) => {
